Refactor and cleanup config of proxy server
