--- config/file-server.json
@@ -1,3 +1,0 @@
-{
-  "port": 9876
-}
