+++ lib/threatconnect.js
@@ -1,11 +1,14 @@
+/*
+ * Copyright (c) 2023, Polarity.io, Inc.
+ */
+
 const querystring = require('querystring');
 const crypto = require('crypto');
 const url = require('url');
 const async = require('async');
 const fp = require('lodash/fp');
 const NodeCache = require('node-cache');
-const { logging } = require('./config/config');
-const fs = require('fs');
+const { logging } = require('../config/config');
 
 const playbookCache = new NodeCache({
   stdTTL: 10 * 60
