+++ styles/threatconnect.less
@@ -28,6 +28,20 @@ a.tc-link-tag {
   margin-bottom: 15px;
 }
 
+.btn-download {
+  border: 1px solid #777;
+  font-weight: normal;
+  font-size: 12px;
+  padding: 0 5px;
+  height: 26px;
+  margin-top: 4px;
+  color: #777;
+
+  &:hover{
+    background-color: #eee;
+  }
+}
+
 .data-block-container:last-child {
   margin-bottom: 0px;
 }
